+++ src/OpenTelemetry/CHANGELOG.md
@@ -11,6 +11,10 @@ Notes](../../RELEASENOTES.md).
   buffer estimate. Such content is now truncated.
   ([#7543](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7543))
 
+* Fixed activity creation throwing when multiple tracer providers return a
+  sampler attribute with the same key.
+  ([#7558](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7558))
+
 ## 1.17.0
 
 Released 2026-Jul-16